LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.meta/script-src-self/worker-import.https-expected.txt

Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=7905a384-4879-4e71-bb44-fc914d98f625&path=/mixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=f69bc9c9-4067-4b9a-ad01-51f930a3ae50&path=/mixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=a1f1d889-70a9-463d-9fba-a246cd4beada&path=/mixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?action=purge&key=be9758c9-79d9-483b-9ae2-6f2640cfe01f&path=%2Fmixed-content

Harness Error (TIMEOUT), message = null

PASS Content Security Policy: Expects allowed for worker-import to same-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects allowed for worker-import to same-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worker-import to same-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects allowed for worker-import to same-https origin and no-redirect redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worker-import to cross-https origin and keep-origin redirection from https context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import to cross-https origin and no-redirect redirection from https context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import to cross-https origin and swap-origin redirection from https context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
TIMEOUT Content Security Policy: Expects blocked for worker-import to same-https origin and swap-origin redirection from https context. Test timed out
NOTRUN Content Security Policy: Expects blocked for worker-import to same-https origin and swap-origin redirection from https context.: securitypolicyviolation
